requirements.txt
flask>=2.0.0
zeroconf>=0.58.0
requests>=2.28.0

# Image / ML
torch>=2.0.0
transformers>=4.30.0
opencv-python>=4.7.0
Pillow>=9.0.0
numpy>=1.24.0

# OCR and utilities
easyocr>=1.6
googletrans==4.0.0-rc1

# Text-to-speech (optional)
gTTS>=2.3.0
playsound==1.2.2
